Edit orthographe mistakes

+++ database/migrations/2020_12_02_133253_add_slug_to_subscription.php
@@ -13,9 +13,11 @@ class AddSlugToSubscription extends Migration
      */
     public function up()
     {
+      if (!Schema::hasTable('subscriptions')) {
         Schema::table('subscriptions', function (Blueprint $table) {
             $table->string('slug')->nullable();
         });
+      }
     }
 
     /**
@@ -25,8 +27,11 @@ class AddSlugToSubscription extends Migration
      */
     public function down()
     {
-        Schema::table('subscriptions', function (Blueprint $table) {
+        if (
+          Schema::hasTable('subscriptions')) {
+          Schema::table('subscriptions', function (Blueprint $table) {
             $table->dropColumn(['slug']);
-        });
+          });
+        }
     }
 }
